tx-generator: Use generic JSON instances
The old manual instances are still available
in the cli with the `legacy-json` command.
