+++ bench/tx-generator/src/Cardano/Benchmarking/Command.hs
@@ -19,12 +19,13 @@ import           Ouroboros.Network.NodeToClient (withIOManager)
 
 import           Cardano.Benchmarking.CliArgsScript
                    (GeneratorCmd, parseGeneratorCmd, runPlainOldCliScript, runEraTransitionTest)
-import           Cardano.Benchmarking.Script (runScript, parseScriptFile)
+import           Cardano.Benchmarking.Script (runScript, parseScriptFileAeson, parseScriptFileLegacy)
 
 data Command
   = CliArguments !GeneratorCmd
   | EraTransition !GeneratorCmd
   | Json !FilePath
+  | LegacyJson !FilePath  
 
 runCommand :: IO ()
 runCommand = withIOManager $ \iocp -> do
@@ -35,9 +36,12 @@ runCommand = withIOManager $ \iocp -> do
     CliArguments   args -> runPlainOldCliScript iocp args >>= handleError
     EraTransition args -> runEraTransitionTest iocp args >>= handleError
     Json file     -> do
-      script <- parseScriptFile file
+      script <- parseScriptFileAeson file
       runScript script iocp >>= handleError
- where
+    LegacyJson file     -> do
+      script <- parseScriptFileLegacy file
+      runScript script iocp >>= handleError
+  where
   handleError :: Show a => Either a b -> IO ()
   handleError = \case
     Right _  -> exitSuccess
@@ -49,6 +53,7 @@ commandParser
     (  cliArgumentsCmd
     <> eraTransitionCmd
     <> jsonCmd
+    <> legacyJsonCmd    
     )
  where
   cliArgumentsCmd = command "cliArguments"
@@ -72,5 +77,13 @@ commandParser
       (  progDesc "tx-generator run JsonScript"
       <> fullDesc
       <> header "tx-generator - run a generic benchmarking script"
+      )
+    )
+
+  legacyJsonCmd = command "legacy-json"
+    (LegacyJson <$> info (strArgument (metavar "FILEPATH"))
+      (  progDesc "tx-generator run JsonScript (legacy format)"
+      <> fullDesc
+      <> header "tx-generator - run a generic benchmarking script (legacy JSON format)"
       )
     )
