+++ bench/tx-generator/tx-generator.cabal
@@ -1,5 +1,5 @@
 name:                  tx-generator
-version:               1.30
+version:               1.32
 description:           The transaction generator for cardano node
 author:                IOHK
 maintainer:            operations@iohk.io
@@ -44,6 +44,7 @@ library
                        Cardano.Benchmarking.Script
                        Cardano.Benchmarking.Script.Action
                        Cardano.Benchmarking.Script.Aeson
+                       Cardano.Benchmarking.Script.AesonLegacy                       
                        Cardano.Benchmarking.Script.Core
                        Cardano.Benchmarking.Script.Env
                        Cardano.Benchmarking.Script.Example
